changing the type of ‘retry_interval’ field of rsh_config_t

+++ rsh.h
@@ -22,7 +22,7 @@ typedef struct {
   char ip[16];
   uint16_t port;
   uint16_t retry_count;
-  uint8_t retry_interval;
+  uint16_t retry_interval;
 } rsh_cfg_t;
 
 #endif  // !RSH_H_
